Windows: deadlock bet. DirEntry lock + DirectoryNodeHdr.TreeLock
The DirectoryNodeHdr.TreeLock must be obtained before the
DirEntry->NonPaged->Lock.  In AFSLocateNameEntry(), the
DirEntry lock is obtained before the TreeLock when processing
a symlink object.  For that case obtain the TreeLOCK first.
Drop it if it is not required.

Reviewed-on: http://gerrit.openafs.org/6558
Tested-by: Jeffrey Altman <jaltman@secure-endpoints.com>
Reviewed-by: Jeffrey Altman <jaltman@secure-endpoints.com>
(cherry picked from commit d9884a480c96d14a3e5a6b08cde5e5003d7a4694)

Change-Id: I7d0226457eafb42870988328f91b3f2ec9ed6e2a
Reviewed-on: http://gerrit.openafs.org/6585
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@secure-endpoints.com>